feat: move landing to root, reorder datasets/collections
