+++ frontend/src/components/Layout/components/defaultLayout.tsx
@@ -1,5 +1,6 @@
 import { useRouter } from "next/router";
 import { FC } from "react";
+import { ROUTES } from "src/common/constants/routes";
 import Layout from "src/components/Layout";
 import { DefaultMainWrapper } from "../style";
 
@@ -7,8 +8,7 @@ const DefaultLayout: FC = ({ children }): JSX.Element => {
   const { pathname } = useRouter();
 
   // REMOVES DEFAULT MAIN WRAPPER ON LANDING PAGE WHICH DISABLES STICKY NAV
-  // CHANGE TO "/" FOR PROD
-  if (pathname === "/landing-page") {
+  if (pathname === ROUTES.HOMEPAGE) {
     return (
       <Layout>
         <main>{children}</main>
