+++ frontend/src/pages/datasets/index.tsx
@@ -1,0 +1,8 @@
+import SidebarLayout from "src/components/Layout/components/sidebarLayout";
+import Datasets from "src/views/Datasets";
+
+const Page = (): JSX.Element => <Datasets />;
+
+Page.Layout = SidebarLayout;
+
+export default Page;
