+++ frontend/src/views/Landing/icons/cellxgene-icon-small.tsx
