+++ frontend/src/views/Landing/icons/expedite-collaboration.tsx
