+++ frontend/src/views/Landing/icons/external-link-arrow.tsx
