+++ frontend/src/views/Landing/icons/gene-expression.tsx
