+++ frontend/src/views/Landing/icons/single-cell-data-active.tsx
